Refactor agent settings pages
