--- frontend/src/app/agents_settings/agent_writers/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function Page() {
-  redirect("/agents_settings/page-writer");
-}
